changed services dir to /lib instead /usr/lib

+++ stage4/02-extras/00-run.sh
@@ -7,8 +7,15 @@ install -v -o 1000 -g 1000 -d "${ROOTFS_DIR}/home/pi/.local/share/applications"
 install -v -o 1000 -g 1000 -d "${ROOTFS_DIR}/home/pi/.local/share/desktop-directories"
 
 # enable deconz-gui autostart
-unlink "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/deconz.service"
-ln -s "/etc/systemd/system/deconz-gui.service" "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/deconz-gui.service"
+if [ -f "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/deconz.service" ]; then
+    unlink "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/deconz.service"
+fi
+
+if [ ! -f "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/deconz-gui.service" ]; then
+    ln -s "/lib/systemd/system/deconz-gui.service" "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/deconz-gui.service"
+fi
 
 # enable vncserver autostart
-ln -s "/usr/lib/systemd/system/vncserver-x11-serviced.service" "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/vncserver-x11-serviced.service"
+if [ ! -f "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/vncserver-x11-serviced.service" ]; then
+    ln -s "/usr/lib/systemd/system/vncserver-x11-serviced.service" "${ROOTFS_DIR}/etc/systemd/system/multi-user.target.wants/vncserver-x11-serviced.service"
+fi
